use NA if there is no pos
Previously if there was no pos, the algorithm would return the maximum sample size, which is incorrect (although negligible in most cases studied so far). To use NA value in Rcpp I had to change the numericVector into IntegerVector in the function simulate_pos.

Overall this should lead to more precise estimates and the output values are more consistent (e.g. NA is shown instead of maximum sample size if there is no pos). The quantiles now ignore missing values.

The tests had to be updated, although the change in the critical pos is minimal (and the values are smaller, as expected).