fix interrupt problem
If you press ESC in RStudio, it will interrupt a C++ call and return nothing. If you press ESC and it happens to be at the time point when Progress::check_abort() is executed, C++ will return a value, handed to an R function. In this case, now, R will return a quite error message and stop. This looks almost the same as the direct interruption in R.

It is not necessary anymore to check at every iteration. But it is still necessary to check every second or so, in case someone is using R and not RStudio.